layer upgrade directory restructure for consistency; minor updates to ui
