+++ nav-app/src/app/layer-upgrade/changelog-cell/changelog-cell.component.ts
