+++ nav-app/src/app/layer-upgrade/layer-upgrade.component.scss
@@ -1,4 +1,4 @@
-@import "../../matrix/matrix-common.scss";
+@import "../matrix/matrix-common.scss";
 
 .title {
     h2 {
